move test to package and use openZeppelin contracts
