+++ contracts/test/token/starkgate/behavior/ERC20.ts
@@ -3,6 +3,7 @@ import { StarknetContract, Account } from 'hardhat/types/runtime'
 import { uint256 } from 'starknet'
 import { toBN } from 'starknet/utils/number'
 import { TIMEOUT } from '../../../constants'
+import { assertErrorMsg } from '../../../utils'
 
 export type TestData = {
   token: StarknetContract
@@ -100,6 +101,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
 
       try {
@@ -110,6 +112,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
     })
 
@@ -123,6 +126,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         })
         throw new Error('This should not pass!')
       } catch (error: any) {
+        console.log(error.message)
         expect(/assert/gi.test(error.message)).to.be.true
       }
 
@@ -191,6 +195,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
       try {
         await t.alice.invoke(t.token, 'transfer', {
@@ -200,6 +205,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
     })
 
@@ -266,6 +272,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
       try {
         await t.owner.invoke(t.token, 'transferFrom', {
@@ -276,6 +283,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
     })
 
@@ -376,6 +384,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
       try {
         await t.owner.invoke(t.token, 'transferFrom', {
@@ -386,6 +395,7 @@ export const shouldBehaveLikeStarkGateERC20 = (beforeFn: BeforeFn) => {
         throw new Error('This should not pass!')
       } catch (error: any) {
         expect(/assert/gi.test(error.message)).to.be.true
+        assertErrorMsg(error?.message, 'SafeUint256: subtraction overflow')
       }
     })
   })
