+++ packages-ts/integration-starkgate/package.json
@@ -18,12 +18,13 @@
     "typescript": "^4.7.2"
   },
   "scripts": {
-    "test": "npx hardhat test --starknet-network devnet"
+    "test": "npx hardhat --network localhost test"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@chainlink-dev/starkgate-contracts": "^0.0.3"
+    "@chainlink-dev/starkgate-contracts": "^0.0.3",
+    "@openzeppelin/contracts": "^4.7.3"
   }
 }
