+++ packages-ts/integration-starkgate/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+        "outDir": "dist",
+        "rootDir": "src"
+    },
+    "include": [
+        "src/**/*"
+    ],
+    "exclude": [
+        "dist",
+        "**/*.spec.ts",
+        "**/*.test.ts"
+    ]
+}
